% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sample_data.R
\name{cms_as_longCms}
\alias{cms_as_longCms}
\title{A phylosql Function}
\usage{
cms_as_longCms(data)
}
\arguments{
\item{data}{data to format}
}
\description{
formats cms data to long for upload. Strict mode.
}
