% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{construct_asv_table}
\alias{construct_asv_table}
\title{A phylosql Function to construct an asv table}
\usage{
construct_asv_table(asv_long)
}
\description{
A phylosql Function to construct an asv table
}
