% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{create_sampleInfo_table}
\alias{create_sampleInfo_table}
\title{A phylosql Function}
\usage{
create_sampleInfo_table(si_long, ...)
}
\arguments{
\item{si_long}{data to upload}
}
\description{
A phylosql Function
}
