% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{fetch_asv_table}
\alias{fetch_asv_table}
\title{A phylosql Function}
\usage{
fetch_asv_table(
  con = NULL,
  database = "eukaryota_sv",
  phylo = FALSE,
  whichSamples = NULL
)
}
\arguments{
\item{con}{connection}

\item{database}{database to send data}

\item{phylo}{logical}

\item{whichSamples}{select specific samples to access}
}
\description{
A phylosql Function
}
