% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{fetch_asvs_by_sample}
\alias{fetch_asvs_by_sample}
\title{A phylosql Function}
\usage{
fetch_asvs_by_sample(
  samples,
  database = NULL,
  con = NULL,
  col = "MetagenNumber"
)
}
\description{
A phylosql Function
}
