% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{fetch_phyloseq}
\alias{fetch_phyloseq}
\title{A phylosql Function}
\usage{
fetch_phyloseq(target_group = NULL, con = get_mtgn_connection())
}
\arguments{
\item{target_group}{target group of organisms. Either amplicon or other species count.}
}
\description{
A phylosql Function
}
