% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{fetch_taxonomy}
\alias{fetch_taxonomy}
\title{A phylosql Function}
\usage{
fetch_taxonomy(
  con = NULL,
  database = "eukaryota_tax",
  whichTaxa = NULL,
  phylo = FALSE
)
}
\arguments{
\item{con}{connection}

\item{database}{database to send data}

\item{whichTaxa}{select specific taxa}

\item{phylo}{logical. Whether to format data for phyloseq or sparseHDD.}
}
\description{
A phylosql Function
}
