% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sample_data.R
\name{fix_soil_moisture}
\alias{fix_soil_moisture}
\title{A phylosql Function}
\usage{
fix_soil_moisture(labdata)

fix_soil_moisture(labdata)
}
\arguments{
\item{labdata}{data to format}
}
\description{
formats lab to for sql upload

formats misc data to cms data
}
