% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sample_data.R
\name{ms_as_cmsData}
\alias{ms_as_cmsData}
\title{A phylosql Function}
\usage{
ms_as_cmsData(data)
}
\arguments{
\item{data}{data to format}
}
\description{
converts mastersheet data to cms data for sql upload
}
