% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sample_data.R
\name{read_masterSheet}
\alias{read_masterSheet}
\title{A phylosql Function}
\usage{
read_masterSheet(
 
    lab_path = "C:/Users/Chris/Metagen/Lab - master sample sheet/master_sample_sheet_v3.0.xlsx"
)
}
\arguments{
\item{lab_path}{path to lab data}
}
\description{
A phylosql Function
}
