% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sample_data.R
\name{transform_dna_yield_to_kg}
\alias{transform_dna_yield_to_kg}
\title{A phylosql Function}
\usage{
transform_dna_yield_to_kg(dna, dna_elution_volume = 2000)
}
\arguments{
\item{dna}{data to format}
}
\description{
formats misc data to cms data
}
