% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_cms}
\alias{update_cms}
\title{A phylosql Function}
\usage{
update_cms(newdata, database = NULL, con = NULL)
}
\arguments{
\item{newdata}{data to upload}

\item{database}{database to send data}

\item{con}{connection}
}
\description{
function to update CMS data
}
