% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_sv}
\alias{update_sv}
\title{A phylosql Function}
\usage{
update_sv(newdata, database, con)
}
\arguments{
\item{newdata}{data to upload}

\item{database}{database to send data}

\item{con}{connection}
}
\description{
function to update SV data
}
