% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{upload_cms_data_Long}
\alias{upload_cms_data_Long}
\title{A phylosql Function}
\usage{
upload_cms_data_Long(data, database = "cmsdatalong", con = NULL)
}
\arguments{
\item{data}{data to upload}

\item{database}{database to send data}

\item{con}{connection}
}
\description{
function to upload cms data to mysql database
}
