% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{upload_lab_data}
\alias{upload_lab_data}
\title{A phylosql Function}
\usage{
upload_lab_data(data, database = "labdata", con = NULL)
}
\arguments{
\item{data}{data to upload}

\item{database}{database to send data}

\item{con}{connection}
}
\description{
function to upload lab data to mysql database
}
