% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{change_metagen_number}
\alias{change_metagen_number}
\title{A phylosql Function}
\usage{
change_metagen_number(
  old_names,
  new_names,
  con,
  databases = list("bacteria_sv", "eukaryota_sv", "labdata", "cmsdata")
)
}
\arguments{
\item{old_names}{data to upload}

\item{new_names}{data to upload}

\item{con}{connection}

\item{database}{database to send data}
}
\description{
function to update lab data
}
