% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{delete_data_by_sample_custom}
\alias{delete_data_by_sample_custom}
\title{A phylosql Function}
\usage{
delete_data_by_sample_custom(samples, database = NULL, con = NULL, col = NULL)
}
\arguments{
\item{database}{database to send data}

\item{con}{connection}

\item{data}{data to upload}
}
\description{
function to delete to mysql database
}
