% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glom_queries.R
\name{sql_phyloseq_by_sample_and_tax_glom}
\alias{sql_phyloseq_by_sample_and_tax_glom}
\title{A phylosql Function. this function is unfinished. Do not use.}
\usage{
sql_phyloseq_by_sample_and_tax_glom(taxa_group, taxalevel, samples, con = NULL)
}
\description{
A phylosql Function. this function is unfinished. Do not use.
}
