% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{upload_bulk_sv}
\alias{upload_bulk_sv}
\title{A phylosql Function}
\usage{
upload_bulk_sv(data, database = NULL, con = NULL)
}
\arguments{
\item{data}{data to upload}

\item{database}{database to send data}

\item{con}{connection}
}
\description{
function to upload a long format SV table to mysql database (quickly)
}
