% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{upload_taxonomy}
\alias{upload_taxonomy}
\title{A phylosql Function}
\usage{
upload_taxonomy(data, database = NULL, con = NULL)
}
\arguments{
\item{data}{data to upload}

\item{database}{database to send data}

\item{con}{connection}
}
\description{
function to upload taxonomy table to mysql database
}
