% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{uploadData}
\alias{uploadData}
\title{A phylosql Function}
\usage{
uploadData(data, tableName, con)
}
\arguments{
\item{data}{data to upload}

\item{tableName}{database to send data}

\item{con}{connection}
}
\description{
A backend function for bulk uploading data to a mysql database
}
